requirements.txt
fastapi==0.121.1
fastjsonschema==2.21.2
filelock==3.20.1
ipython==8.12.3
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
load-dotenv==0.1.0
matplotlib-inline==0.2.1
numpy==2.2.6
pandas==2.3.3
psycopg2==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
stack-data==0.6.3
sqlalchemy
tqdm==4.67.1
uvicorn==0.38.0
xgboost
lightgbm
optuna
mlflow
shap
dagshub